+++ interactions/ext/voice/__init__.py
@@ -1,4 +1,1 @@
-from .client import *  # noqa: F401 F403
-from .enums import *  # noqa: F401 F403
-from .error import *  # noqa: F401 F403
-from .voice import *  # noqa: F401 F403
+from .client import VoiceClient
