+++ pyproject.toml
@@ -1,0 +1,31 @@
+[tool.poetry]
+name = "interactions-voice"
+version = "1.0.0"
+description = "A voice-capable client for interactions.py"
+authors = ["EdVraz <edvraz12@gmail.com>"]
+readme = "README.md"
+repository = "https://github.com/interactions-py/voice"
+classifiers=[
+    "Programming Language :: Python :: 3",
+    "Development Status :: 4 - Beta",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+]
+packages = [
+    { include = "interactions/ext/voice" },
+]
+license = "MIT"
+
+[tool.poetry.dependencies]
+python = "^3.8.6"
+PyNaCl = "^1.5.0"
+discord-py-interactions = {git = "https://github.com/interactions-py/library.git", rev = "4.1.1-rc.1"}
+
+[tool.poetry.dev-dependencies]
+black = "^22.3.0"
+isort = "^5.10.1"
+pyright = "^1.1.235"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
